Add EPA gui
