+++ anise-gui/src/ui.rs
@@ -1,14 +1,13 @@
 use anise::{almanac::Almanac, errors::AlmanacError};
 use eframe::egui;
 use egui::Theme;
-use egui_extras::{Column, TableBuilder};
 use hifitime::TimeScale;
 
 use log::error;
 #[cfg(target_arch = "wasm32")]
 use poll_promise::Promise;
 
-use crate::{bpc::bpc_ui, spk::spk_ui};
+use crate::{bpc::bpc_ui, epa::epa_ui, pca::pca_ui, spk::spk_ui};
 
 #[cfg(target_arch = "wasm32")]
 type AlmanacFile = Option<(String, Vec<u8>)>;
@@ -128,7 +127,8 @@ impl eframe::App for UiApp {
                         match &self.path {
                             None => {
                                 let mut trigger_file_load = false;
-                                trigger_file_load |= ui.button("Select file to inspect...").clicked();
+                                trigger_file_load |=
+                                    ui.button("Select file to inspect...").clicked();
 
                                 // If we are in the browser, we need to also check if the file
                                 // is ready to be loaded instead of just checking if the button
@@ -140,18 +140,17 @@ impl eframe::App for UiApp {
 
                                 // Show the open file dialog
                                 if trigger_file_load {
-                                        // Try to load this file
-                                        match self.load_almanac() {
-                                            FileLoadResult::NoFileSelectedYet => {
-                                            }
-                                            FileLoadResult::Ok((path, almanac)) => {
-                                                self.almanac = almanac;
-                                                self.path = Some(path);
-                                            }
-                                            FileLoadResult::Error(e) => {
-                                                error!("{e}");
-                                            }
+                                    // Try to load this file
+                                    match self.load_almanac() {
+                                        FileLoadResult::NoFileSelectedYet => {}
+                                        FileLoadResult::Ok((path, almanac)) => {
+                                            self.almanac = almanac;
+                                            self.path = Some(path);
                                         }
+                                        FileLoadResult::Error(e) => {
+                                            error!("{e}");
+                                        }
+                                    }
                                 }
                             }
                             Some(path) => {
@@ -178,7 +177,7 @@ impl eframe::App for UiApp {
 
                                 let mut unload_file = false;
                                 ui.vertical(|ui| {
-                                    ui.horizontal(|ui|{
+                                    ui.horizontal(|ui| {
                                         ui.label(format!("Inspecting {path}"));
                                         if ui.button("Close").clicked() {
                                             unload_file = true;
@@ -192,11 +191,21 @@ impl eframe::App for UiApp {
                                         ui.text_edit_singleline(&mut format!("{crc}"));
 
                                         if label.ends_with("SPK") {
-                                            let num_summaries = self.almanac.spk_data[0].as_ref().unwrap().daf_summary().unwrap().num_summaries();
+                                            let num_summaries = self.almanac.spk_data[0]
+                                                .as_ref()
+                                                .unwrap()
+                                                .daf_summary()
+                                                .unwrap()
+                                                .num_summaries();
                                             ui.label("Number of summaries");
                                             ui.label(format!("{num_summaries}"));
                                         } else if label.ends_with("PCK") {
-                                            let num_summaries = self.almanac.bpc_data[0].as_ref().unwrap().daf_summary().unwrap().num_summaries();
+                                            let num_summaries = self.almanac.bpc_data[0]
+                                                .as_ref()
+                                                .unwrap()
+                                                .daf_summary()
+                                                .unwrap()
+                                                .num_summaries();
                                             ui.label("Number of summaries");
                                             ui.label(format!("{num_summaries}"));
                                         }
@@ -226,176 +235,29 @@ impl eframe::App for UiApp {
                                                     }
                                                 });
 
-                                            ui.checkbox(
-                                                &mut self.show_unix,
-                                                "UNIX timestamps",
-                                            );
+                                            ui.checkbox(&mut self.show_unix, "UNIX timestamps");
                                         });
                                     }
 
                                     // Now display the data
                                     if label == "DAF/PCK" {
-                                        bpc_ui(ui, &self.almanac, self.show_unix, self.selected_time_scale);
+                                        bpc_ui(
+                                            ui,
+                                            &self.almanac,
+                                            self.show_unix,
+                                            self.selected_time_scale,
+                                        );
                                     } else if label == "DAF/SPK" {
-                                        spk_ui(ui, &self.almanac, self.show_unix, self.selected_time_scale)
+                                        spk_ui(
+                                            ui,
+                                            &self.almanac,
+                                            self.show_unix,
+                                            self.selected_time_scale,
+                                        );
                                     } else if label == "ANISE/PCA" {
-                                        TableBuilder::new(ui)
-                                            .column(Column::auto().at_least(100.0).resizable(true))
-                                            .column(Column::auto().at_least(50.0).resizable(true))
-                                            .column(Column::auto().at_least(75.0).resizable(true))
-                                            .column(Column::auto().at_least(75.0).resizable(true))
-                                            .column(Column::auto().at_least(75.0).resizable(true))
-                                            .column(Column::auto().at_least(125.0).resizable(true))
-                                            .column(Column::auto().at_least(125.0).resizable(true))
-                                            .column(Column::auto().at_least(125.0).resizable(true))
-                                            .column(Column::remainder())
-                                            .header(20.0, |mut header| {
-                                                header.col(|ui| {
-                                                    ui.heading("Name");
-                                                });
-                                                header.col(|ui| {
-                                                    ui.heading("ID");
-                                                });
-                                                header.col(|ui| {
-                                                    ui.heading("Gravity param (km^3/s^2)");
-                                                });
-
-                                                header.col(|ui| {
-                                                    ui.heading("Major axis (km)");
-                                                });
-                                                header.col(|ui| {
-                                                    ui.heading("Minor axis (km)");
-                                                });
-                                                header.col(|ui| {
-                                                    ui.heading("Polar axis (km)");
-                                                });
-
-                                                header.col(|ui| {
-                                                    ui.heading("Pole right asc.");
-                                                });
-                                                header.col(|ui| {
-                                                    ui.heading("Pole declination");
-                                                });
-                                                header.col(|ui| {
-                                                    ui.heading("Prime meridian");
-                                                });
-                                            })
-                                            .body(|mut body| {
-                                                let pck = &self.almanac.planetary_data;
-
-                                                let binding = pck.lut.entries();
-                                                let mut values = binding.values().collect::<Vec<_>>().to_vec();
-                                                values.sort_by_key(|(opt_id, _)| match opt_id {
-                                                    Some(id) => *id,
-                                                    None => 0
-                                                });
-
-                                                for (opt_id, opt_name) in values
-                                                {
-                                                    let data = if let Some(id) = opt_id {
-                                                        pck.get_by_id(*id).unwrap()
-                                                    } else {
-                                                        pck.get_by_name(&opt_name.clone().unwrap()).unwrap()
-                                                    };
-
-                                                    body.row(30.0, |mut row| {
-                                                        row.col(|ui| {
-                                                            ui.label(match opt_name {
-                                                                Some(name) => format!("{name}"),
-                                                                None => "Unset".to_string(),
-                                                            });
-                                                        });
-
-                                                        row.col(|ui| {
-                                                            ui.label(match opt_id {
-                                                                Some(id) => format!("{id}"),
-                                                                None => "Unset".to_string(),
-                                                            });
-                                                        });
-
-                                                        row.col(|ui| {
-                                                            ui.text_edit_singleline(&mut format!(
-                                                                "{}",
-                                                                data.mu_km3_s2
-                                                            ));
-                                                        });
-
-                                                        match data.shape {
-                                                            None => {
-                                                                // Three empty columns
-                                                                row.col(|ui| {
-                                                                    ui.label("Unset");
-                                                                });
-                                                                row.col(|ui| {
-                                                                    ui.label("Unset");
-                                                                });
-                                                                row.col(|ui| {
-                                                                    ui.label("Unset");
-                                                                });
-                                                            }
-                                                            Some(shape) => {
-                                                                row.col(|ui| {
-                                                                    ui.text_edit_singleline(
-                                                                        &mut format!(
-                                                                            "{}",
-                                                                            shape.semi_major_equatorial_radius_km
-                                                                        ),
-                                                                    );
-                                                                });
-                                                                row.col(|ui| {
-                                                                    ui.text_edit_singleline(
-                                                                        &mut format!(
-                                                                            "{}",
-                                                                            shape.semi_minor_equatorial_radius_km
-                                                                        ),
-                                                                    );
-                                                                });
-                                                                row.col(|ui| {
-                                                                    ui.text_edit_singleline(
-                                                                        &mut format!(
-                                                                            "{}",
-                                                                            shape.polar_radius_km
-                                                                        ),
-                                                                    );
-                                                                });
-                                                            }
-                                                        }
-
-                                                        match data.pole_right_ascension {
-                                                            None => row.col(|ui| {
-                                                                ui.label("Unset");
-                                                            }),
-                                                            Some(pole_ra) => {
-                                                                row.col(|ui| {
-                                                                    ui.label(format!("{pole_ra}"));
-                                                                })
-                                                            }
-                                                        };
-
-                                                        match data.pole_declination {
-                                                            None => row.col(|ui| {
-                                                                ui.label("Unset");
-                                                            }),
-                                                            Some(pole_dec) => {
-                                                                row.col(|ui| {
-                                                                    ui.label(format!("{pole_dec}"));
-                                                                })
-                                                            }
-                                                        };
-
-                                                        match data.prime_meridian {
-                                                            None => row.col(|ui| {
-                                                                ui.label("Unset");
-                                                            }),
-                                                            Some(pm) => {
-                                                                row.col(|ui| {
-                                                                    ui.label(format!("{pm}"));
-                                                                })
-                                                            }
-                                                        };
-                                                    });
-                                                }
-                                            });
+                                        pca_ui(ui, &self.almanac);
+                                    } else if label == "ANISE/EPA" {
+                                        epa_ui(ui, &self.almanac);
                                     }
                                 });
